# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movie API
  version: 1.0.0
servers:
- url: https://localhost:8080/api/v1
...
paths:
...
  /promotions:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Promotion'
                type: array
          description: OK
      summary: List all promotions
      tags:
      - Promotions
      x-accepts: application/json
      x-tags:
      - tag: Promotions
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Promotion'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
          description: Created
      summary: Create a new promotion
      tags:
      - Promotions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Promotions
  /promotions/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: Delete a promotion
      tags:
      - Promotions
      x-accepts: application/json
      x-tags:
      - tag: Promotions
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
          description: OK
        "404":
          description: Not Found
      summary: Get details of a specific promotion
      tags:
      - Promotions
      x-accepts: application/json
      x-tags:
      - tag: Promotions
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Promotion'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update an existing promotion
      tags:
      - Promotions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Promotions
...
components:
...
  schemas:
...
    Promotion:
      example:
        discountPercent: 0.2
        endDate: 2024-08-31T23:59:59Z
        name: Summer Sale
        description: Get 20% off all items!
        promotionId: 1
        startDate: 2024-07-15T00:00:00Z
      properties:
        promotionId:
          description: Unique identifier for the promotion
          example: 1
          type: integer
        name:
          description: Name of the promotion
          example: Summer Sale
          type: string
        description:
          description: Description of the promotion
          example: Get 20% off all items!
          type: string
        discountPercent:
          description: Discount percentage offered by the promotion (0-1)
          example: 0.2
          type: number
        startDate:
          description: Start date of the promotion (ISO 8601 format)
          example: 2024-07-15T00:00:00Z
          format: date-time
          type: string
        endDate:
          description: End date of the promotion (ISO 8601 format)
          example: 2024-08-31T23:59:59Z
          format: date-time
          type: string
      required:
      - description
      - discountPercent
      - endDate
      - name
      - promotionId
      - startDate
      type: object
...